StIcon: only compute shadow pipeline when the texture is properly allocated
Creating the shadow pipeline requires the actor to be allocated in order
to get its dimensions, however in the current state we just compute it
even if this is not the case.
This causes _st_create_shadow_pipeline_from_actor (when getting the allocation
box) to trigger an allocation cycle, which might lead to a convolution to
st_icon_finish_update causing breakage on data as soon as we return from it.

Waiting for the texture size change before trying to update the shadow pipeline
is a way for avoiding this.

https://bugzilla.gnome.org/show_bug.cgi?id=788908